refactor: split QA in separate modules
Only files have been moved. No code has been changed.
