+++ qa/abstracts/src/test/resources/automated-tests.bpmn
