+++ qa/abstracts/src/test/resources/call-activity.bpmn
