+++ qa/abstracts/src/test/resources/looping-servicetask.bpmn
