+++ qa/abstracts/src/test/resources/message-event.bpmn
