+++ qa/abstracts/src/test/resources/message-start-event.bpmn
