+++ qa/abstracts/src/test/resources/start-end.bpmn
