+++ qa/abstracts/src/test/resources/timer-start-event-daily.bpmn
