--- qa/src/test/java/io/camunda/zeebe/process/test/qa/abstracts/assertions/AbstractProcessAssertTest.java
@@ -1,257 +1,0 @@
-/*
- * Copyright © 2021 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.zeebe.process.test.qa.abstracts.assertions;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import io.camunda.zeebe.client.ZeebeClient;
-import io.camunda.zeebe.client.api.response.DeploymentEvent;
-import io.camunda.zeebe.process.test.api.ZeebeTestEngine;
-import io.camunda.zeebe.process.test.assertions.BpmnAssert;
-import io.camunda.zeebe.process.test.assertions.ProcessAssert;
-import io.camunda.zeebe.process.test.qa.util.Utilities;
-import io.camunda.zeebe.process.test.qa.util.Utilities.ProcessPackLoopingServiceTask;
-import java.util.concurrent.TimeoutException;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-
-public abstract class AbstractProcessAssertTest {
-
-  public static final String WRONG_VALUE = "wrong value";
-
-  // These tests are for testing assertions as well as examples for users
-  @Nested
-  class HappyPathTests {
-
-    private ZeebeClient client;
-    private ZeebeTestEngine engine;
-
-    @Test
-    void testHasBPMNProcessId() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      processAssert.hasBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-    }
-
-    @Test
-    void testHasVersion() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      processAssert.hasVersion(1);
-    }
-
-    @Test
-    void testHasResourceName() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      processAssert.hasResourceName(ProcessPackLoopingServiceTask.RESOURCE_NAME);
-    }
-
-    @Test
-    void testHasAnyInstances() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      Utilities.startProcessInstance(engine, client, ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      processAssert.hasAnyInstances();
-    }
-
-    @Test
-    void testHasNoInstances() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      processAssert.hasNoInstances();
-    }
-
-    @Test
-    void testHasInstances() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      Utilities.startProcessInstance(engine, client, ProcessPackLoopingServiceTask.PROCESS_ID);
-      Utilities.startProcessInstance(engine, client, ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      processAssert.hasInstances(2);
-    }
-  }
-
-  // These tests are just for assertion testing purposes. These should not be used as examples.
-  @Nested
-  class UnhappyPathTests {
-
-    private ZeebeClient client;
-    private ZeebeTestEngine engine;
-
-    @Test
-    void testHasBPMNProcessIdFailure() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      assertThatThrownBy(() -> processAssert.hasBpmnProcessId(WRONG_VALUE))
-          .isInstanceOf(AssertionError.class)
-          .hasMessage(
-              "Expected BPMN process ID to be '%s' but was '%s' instead.",
-              WRONG_VALUE, ProcessPackLoopingServiceTask.PROCESS_ID);
-    }
-
-    @Test
-    void testHasVersionFailure() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      assertThatThrownBy(() -> processAssert.hasVersion(12345))
-          .isInstanceOf(AssertionError.class)
-          .hasMessage("Expected version to be 12345 but was 1 instead");
-    }
-
-    @Test
-    void testHasResourceNameFailure() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      assertThatThrownBy(() -> processAssert.hasResourceName(WRONG_VALUE))
-          .isInstanceOf(AssertionError.class)
-          .hasMessage(
-              "Expected resource name to be '%s' but was '%s' instead.",
-              WRONG_VALUE, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-    }
-
-    @Test
-    void testHasAnyInstancesFailure() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      assertThatThrownBy(processAssert::hasAnyInstances)
-          .isInstanceOf(AssertionError.class)
-          .hasMessage("The process has no instances");
-    }
-
-    @Test
-    void testHasNoInstances() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      Utilities.startProcessInstance(engine, client, ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      assertThatThrownBy(processAssert::hasNoInstances)
-          .isInstanceOf(AssertionError.class)
-          .hasMessage("The process does have instances");
-    }
-
-    @Test
-    void testHasInstances() throws InterruptedException, TimeoutException {
-      // given
-      final DeploymentEvent deploymentEvent =
-          Utilities.deployProcess(client, ProcessPackLoopingServiceTask.RESOURCE_NAME);
-
-      // when
-      Utilities.startProcessInstance(engine, client, ProcessPackLoopingServiceTask.PROCESS_ID);
-      Utilities.startProcessInstance(engine, client, ProcessPackLoopingServiceTask.PROCESS_ID);
-      Utilities.startProcessInstance(engine, client, ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      final ProcessAssert processAssert =
-          BpmnAssert.assertThat(deploymentEvent)
-              .extractingProcessByBpmnProcessId(ProcessPackLoopingServiceTask.PROCESS_ID);
-
-      // then
-      assertThatThrownBy(() -> processAssert.hasInstances(2))
-          .isInstanceOf(AssertionError.class)
-          .hasMessage("Expected number of instances to be 2 but was 3 instead");
-    }
-  }
-}
