--- qa/src/test/java/io/camunda/zeebe/process/test/qa/abstracts/inspections/AbstractProcessInstanceInspectionsTest.java
@@ -1,85 +1,0 @@
-/*
- * Copyright © 2021 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.zeebe.process.test.qa.abstracts.inspections;
-
-import io.camunda.zeebe.client.ZeebeClient;
-import io.camunda.zeebe.client.api.response.ProcessInstanceEvent;
-import io.camunda.zeebe.process.test.api.ZeebeTestEngine;
-import io.camunda.zeebe.process.test.assertions.BpmnAssert;
-import io.camunda.zeebe.process.test.inspections.InspectionUtility;
-import io.camunda.zeebe.process.test.inspections.model.InspectedProcessInstance;
-import io.camunda.zeebe.process.test.qa.util.Utilities;
-import io.camunda.zeebe.process.test.qa.util.Utilities.ProcessPackCallActivity;
-import java.util.Optional;
-import java.util.concurrent.TimeoutException;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public abstract class AbstractProcessInstanceInspectionsTest {
-
-  @Test
-  void testStartedByProcessInstanceWithProcessId() throws InterruptedException, TimeoutException {
-    // given
-    Utilities.deployProcesses(
-        getClient(),
-        ProcessPackCallActivity.RESOURCE_NAME,
-        ProcessPackCallActivity.CALLED_RESOURCE_NAME);
-    final ProcessInstanceEvent instanceEvent =
-        Utilities.startProcessInstance(
-            getEngine(), getClient(), ProcessPackCallActivity.PROCESS_ID);
-
-    // when
-    final Optional<InspectedProcessInstance> firstProcessInstance =
-        InspectionUtility.findProcessInstances()
-            .withParentProcessInstanceKey(instanceEvent.getProcessInstanceKey())
-            .withBpmnProcessId(ProcessPackCallActivity.CALLED_PROCESS_ID)
-            .findFirstProcessInstance();
-
-    // then
-    Assertions.assertThat(firstProcessInstance).isNotEmpty();
-    BpmnAssert.assertThat(firstProcessInstance.get()).isCompleted();
-    BpmnAssert.assertThat(instanceEvent)
-        .hasPassedElement(ProcessPackCallActivity.CALL_ACTIVITY_ID)
-        .isCompleted();
-  }
-
-  @Test
-  void testStartedByProcessInstanceWithProcessId_wrongId()
-      throws InterruptedException, TimeoutException {
-    // given
-    Utilities.deployProcesses(
-        getClient(),
-        ProcessPackCallActivity.RESOURCE_NAME,
-        ProcessPackCallActivity.CALLED_RESOURCE_NAME);
-    final ProcessInstanceEvent instanceEvent =
-        Utilities.startProcessInstance(
-            getEngine(), getClient(), ProcessPackCallActivity.PROCESS_ID);
-
-    // when
-    final Optional<InspectedProcessInstance> firstProcessInstance =
-        InspectionUtility.findProcessInstances()
-            .withParentProcessInstanceKey(instanceEvent.getProcessInstanceKey())
-            .withBpmnProcessId("wrongId")
-            .findFirstProcessInstance();
-
-    // then
-    Assertions.assertThat(firstProcessInstance).isEmpty();
-  }
-
-  public abstract ZeebeClient getClient();
-
-  public abstract ZeebeTestEngine getEngine();
-}
